refactor: remove more unused code
