--- apps/admin-ui/nginx/nginx.conf
@@ -1,43 +1,0 @@
-worker_processes  1;
-
-error_log  /var/log/nginx/error.log warn;
-pid        /tmp/nginx.pid;
-
-events {
-    worker_connections  1024;
-}
-
-http {
-    include                 /etc/nginx/mime.types;
-    default_type            application/octet-stream;
-    log_format              main  '$remote_addr - $remote_user [$time_local] "$request" '
-                            '$status $body_bytes_sent "$http_referer" '
-                            '"$http_user_agent" "$http_x_forwarded_for"';
-    access_log              /var/log/nginx/access.log  main;
-    sendfile                on;
-    keepalive_timeout       65;
-    client_body_temp_path   /tmp/client_temp;
-    proxy_temp_path         /tmp/proxy_temp_path;
-    fastcgi_temp_path       /tmp/fastcgi_temp;
-    uwsgi_temp_path         /tmp/uwsgi_temp;
-    scgi_temp_path          /tmp/scgi_temp;
-
-    server {
-        listen          8000;
-        server_name     localhost;
-        root            /usr/share/nginx/html;
-        location / {
-            root        /usr/share/nginx/html;
-            try_files   $uri /index.html;
-        }
-        gzip on;
-        gzip_disable "msie6";
-        gzip_vary on;
-        gzip_proxied any;
-        gzip_comp_level 6;
-        gzip_buffers 16 8k;
-        gzip_http_version 1.1;
-        gzip_min_length 0;
-        gzip_types text/plain application/javascript text/css application/json application/x-javascript text/xml application/xml application/xml+rss text/javascript application/vnd.ms-fontobject application/x-font-ttf font/opentype;
-    }
-}
