+++ apps/admin-ui/src/component/my-units/MyUnitRecurringReservation/RecurringReservationDone.tsx
@@ -7,9 +7,9 @@ import { ErrorBoundary } from "react-error-boundary";
 import { z } from "zod";
 import { H1, H6 } from "common/src/common/typography";
 import { breakpoints } from "common/src/common/style";
-import RecurringReservationsView from "app/component/reservations/requested/RecurringReservationsView";
+import RecurringReservationsView from "@/component/reservations/requested/RecurringReservationsView";
 import { ActionsWrapper } from "./commonStyling";
-import ReservationList from "../../ReservationsList";
+import { ReservationList } from "@/component/ReservationsList";
 
 const InfoSection = styled.p`
   margin: var(--spacing-l) 0;
