--- packages/common/src/ExampleCommonComponent.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Common = (): JSX.Element => {
-  return <div>Hello from common!</div>;
-};
-
-export default Common;
